requirements.txt
flask
readability-lxml
requests
gunicorn
